admin page query(nested query) for historical orders function modified, orders status update function finished.

+++ sky-pojo/src/main/java/com/sky/vo/OrderStatisticsVO.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 @Data
 public class OrderStatisticsVO implements Serializable {
     //待接单数量
-    private Integer toBeConfirmed;
+    private Long toBeConfirmed;
 
     //待派送数量
     private Integer confirmed;
